Extract chunk pipeline helper from indexer
